add `header.id_field_name`, header type inside other structures
`header.id_field_name` is needed to tell the generic dissector
plugin which field it needs to choose the correct structure

`build_struct` is reused: to build headers, leave `header_type_name` to
None; to build struct set it to the name of the header.

Temporarily fix builder tests: they'll be soon moved away.
The way Pystache and other mustache implementations use to
test themselves is interesting, I'll need to check if it's ok
for Lelei.