fix short name not printed in wsgd, add non-regression tests
PROTOSHORTNAME and PROTOABBREV were not initialized: it causes
problems during wsgd parsing.

tests for this problem are now added in `test_builder`.
They're designed to check not only the values are initialized,
but if they're initialized with the right values too.
a regex is employed to retrieve the initialization values.

tests about MSG_HEADER_TYPE and MSG_ID_FIELD_NAME are added too.