Add MSG_MAIN_TYPE: how a message is chosen
MSG_MAIN_TYPE is the way the generic dissector handler chooses which
structure it should use to parse the packet.
As we have a single structure, we'll insert the name of the structure.
In the future, when we may let the user insert 1+ structures,
a simple `switch` directive will be added

```
switch T_msg_switch T_type_message
{
  case enum_t::enum_value: structure "";
  ...
  default:
    header_t header;
    raw(*) end_of_message;
}
```
or something like that.