fix lelei not run correctly, uint64/int64 definition problem
to be clear: int64 is the one that exists, not uint64.
a test about it has been added too.
`
solve problems with unrecognized hashbangs: it took a run
of `dos2unix bin/lelei` to fix it. It seems that [1]
^M control chars are responsible for strange errors
during the startup of scripts: the shell cannot
recognize which interpreter has to use, leadin to the error
explained in the SO question.

[1]: http://stackoverflow.com/a/19765741